requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-community
langchain-chroma
langchain-huggingface
langchain-groq
chromadb
sentence-transformers
pypdf
arxiv
wikipedia
validators==0.28.1
youtube_transcript_api